requirements.txt
duckduckgo-search
openai
KSockets
pyinstaller
python-dotenv
pyyaml
loguru